pyproject.toml
[project]
name = "stock-price-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2"
]
